Fix deletion behavior around atomic blocks
Summary:
There is currently a selection state bug for deletion behavior around atomic blocks. Namely, removing an atomic block from a neighboring block will remove the atomic block but leave the cursor in an invalid state.

Make it so that atomic blocks will be removed during:

- Forward deletion from the last offset of a preceding block
- Backward deletion from the zero offset of a following block

tasti and I discussed whether there should be some special behavior for forward deletion from an empty preceding block, i.e. removing the current block and moving the cursor upward, but this seemed magical.

We can try this out and adjust the behavior as needed.
Closes https://github.com/facebook/draft-js/pull/332

Reviewed By: ezequiel

Differential Revision: D3245579

fb-gh-sync-id: a250cf87515f225d13e1da30023b9fb681e06ef5
fbshipit-source-id: a250cf87515f225d13e1da30023b9fb681e06ef5