block level metadata
Summary:
This commit add support for block level metadata.

It is the same feature as https://github.com/facebook/draft-js/pull/157

This PR does not rely on entity, but just add a data field to ContentBlock.

hense block metadata are store in contentState and are supported by undo/redo.

Why not using entities?
- Entity are not made for block. what does type and mutability would map to?
- Entity are planned to be integrated into contentState anyway (see #185)
Closes https://github.com/facebook/draft-js/pull/216

Reviewed By: spicyj

Differential Revision: D3123774

fbshipit-source-id: 27fdf0090d8faafba2848cb72f584987af3bbe0e