removes window.prompt from media example
fixes #463

this fixes two things:
1. it fixes #463 where if you inserted a blank string for a piece
of media the editor state would be set to null.
2. the media example no longer uses window.prompt which previously
broke selection state. instead using an input like the link example.

it should be noted though that with this commit it's possible to
insert a blank url for media. i initially was going to add protection
for this, but after looking at the link example it appears the link
example doesn't care if you enter a blank link.

looking at the media objects generated for mp3/mp4 the play bar is still
there, and image is a whitespace. i'll happily add in a safeguard if
it's determined that's "unwanted" behavior. however I decided to stick
with consistency with the other examples for this commit.