aria-haspop indicates popup menus. It is not needed here
Summary: `aria-haspopup` is used to indicate "popup button"; buttons that, when pressed, show a little menu of options. It's not meant to indicate any element that has a dialog associated with it. In the case of draft.js, the `aria-expanded` attribute is used to indicate that the input has a dropdown with options.

Reviewed By: spicyj

Differential Revision: D5550366

fbshipit-source-id: 90249146a7f69f538ab2177842ee925825238ccb