Update handleKeyCommand examples and docs
Summary:
**Summary**
handleKeyCommand receives a second parameter with the latest instance of the editor state. When this function needs the editor state it should use this instance instead of any other reference.

This update reflects a change made to handleKeyCommand to also receive the editor state. Since it can be potentially changed by Draft during key handling any previous references might be of an old version of the state.

**Test Plan**
Loaded `examples/draft-0-10-0/{media/media.html, rich/rich.html, tex/}` in the browser and verified the text input worked as expected and in particular that we could delete deadkeys like `.
Closes https://github.com/facebook/draft-js/pull/1346

Differential Revision: D5751246

fbshipit-source-id: 63ac45facd21c0c87de983ebfbcf109ae0e486c9