Making CharacterList optional when creating ContentBlock
Summary:
Background:

Relying on the ContentBlock constructor for record creation means that the user would have to create a valid characterList matching the length of the text passed. It makes sense to simplify the API a bit by allowing the characterList to be optional.

This fixes: https://github.com/facebook/draft-js/issues/1455
Closes https://github.com/facebook/draft-js/pull/1472

Differential Revision: D6211343

fbshipit-source-id: a803e891833474b055aed57c14abb78e927ccf32