Refactoring our test suite and introducing jest snapshots
Summary:
This PR aim on refactoring our existing tests and moving towards jest snapshot testing.

Snapshot testing will not only allow us to move faster but also reduce the barrier for contributors to be able to contribute to draft.

tldr:

- moved to jest snapshots
- refactored tests to remove duplication
- removed nested describes in favor of plain 'test' to increase readability

***

I suggest we review this internally in phabricator as it may be tricky to do that on github. Also once this is merged should be pretty easy to update tests or add some new ones.
Closes https://github.com/facebook/draft-js/pull/1490

Reviewed By: flarnie

Differential Revision: D6292507

Pulled By: flarnie

fbshipit-source-id: 915089257fad8aa5e022659f47c60d2ec5ac6c12