Mark props that have defaults as required
Summary:
Fixes facebook/draft-js#1514

**Summary**

According to the [Flow documentation](https://flow.org/en/docs/react/components/#toc-using-default-props), when using default props, the prop should be marked as required, as "Flow will make sure that [the prop] is optional if you have a default prop for [it]."

This fixes Flow errors that would occur when using Draft.js in a project with Flow. Flow would complain that `blockRendererFn` and `blockStyleFn` could potentially be `undefined` when passed to `DraftEditorContents`, even though these two properties have default values, so they will never be undefined.

This commit also upgrades Flow to 0.59.0, since there weren't any additional Flow errors between the version of Flow that was being used (0.53.1) and the latest version of Flow (0.59.0).

**Test Plan**

 * Flow check passes
 * A sample project, such as https://github.com/bjohn465/draft-and-flow-errors, does not produce Flow errors when using Draft.js
Closes https://github.com/facebook/draft-js/pull/1520

Differential Revision: D6366083

fbshipit-source-id: 142ba22cab5d87ccbb65d6090a9634317b9e2dfa