Exploration - Feature tree / Tree adapter transformer
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Tree adapter to transformer**

This creates a raw tree data adapter to transform back and forth between raw tree and draft raw data

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1521

Differential Revision: D6375517

fbshipit-source-id: 548fddb1b56e066b7550d42971b7189ba38feda2