Exploration - Feature tree / Integrating GK flag
Summary:
This PR is part of a series of PR's that will be exploring tree data block support in Draft.

Integrating GK flag

This PR adds support for the feature flag where we will use to conditionally load the DraftEditorContents-core.react.js or DraftEditorContentsExperimental.react.js.

Will import this and have the review done internally since it will require some internal wiring for it to work

Note:
This is unstable and not part of the public API and should not be used by
production systems.

Reviewed By: flarnie

Differential Revision: D6383180

fbshipit-source-id: fec9cd2962b881c74a69569654fbf7e5b2f5fb61