Exploration - Feature tree / Adding support for nested block selection
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for nested block selection**

This PR adds support for dealing with nested blocks selections and also add some tests to make sure we cover all the previous use cases when dealing with nested blocks.

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1529

Differential Revision: D6439187

fbshipit-source-id: 520e50f806eb68018c317cb40d296ca2c9c3a0bd