Exploration - Feature tree / Adding support for container text nodes
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for container text nodes**

This PR will add support for handling container text nodes that precedes real block nodes.

example:

```html
<blockquote>
   some text
   <h1>header block</h1>
</blockquote>
```

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1533

Differential Revision: D6442010

fbshipit-source-id: 70e829ccebf9b503877f201f34e6b1e03593e777