Exploration - Feature tree / Adding support for moving tree blocks
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for moving tree blocks**

This PR adds support for moving tree blocks via the moveBlockInContentState transaction.

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1540

Differential Revision: D6456754

fbshipit-source-id: c802b0069f11f9e6b10734f4465f19edf828522e