Exploration - Feature tree / adding internal support for tree data structures
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for tree data structures**

This PR includes the missing parts for adding support for internal tree data structure by providing:

* last transaction insertFragmentIntoContentState.js
* last modifier AtomicBlockUtils
* fixing a bug on the randomizeBlockMapKeys to handle orphanated fragments

This is the last PR of the series to add support for internal tree data structures, from here we should then either update `RichTextEditorUtil` to have some `tree` awareness and having some controls for creating nesting content or create a new `NestedRichTextEditorUtil` that can be used for that. Ideally that work would also come with a `demo` example page

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1593

Differential Revision: D6679776

fbshipit-source-id: ad25593a7dd01552848a9ae2cdc7d34491b8130e