Refactor `convertFromHTMLtoContentBlocks`
Summary: One of the utility helper methods commonly used to manage state with Draft.js is the `convertFromHTMLtoContentBlocks` method, which converts an HTML string into a data structure used by Draft.js for tracking state. This is a helpful utility, but it was written in an overly complex way. This commits adds a new simpler implementation of `convertFromHTMLToContentBlocks`.

Reviewed By: flarnie

Differential Revision: D7158000

fbshipit-source-id: 3424cdeb03663dc6c5a194da3d317c6ab97507dc