Actually for real fix the flow type of decorator props
Summary:
The 'entityKey' could always have been null/undefined, and it's time that the
flow type reflects that. Our code should handle the case, though unexpected,
when it could be undefined.

A previous diff[1] was submitted but somehow the changes were not all commited.
This finishes the job, updating decorator components in our product code and the
type definition in Draft.js.

[1]: https://our.intern.facebook.com/intern/diff/D7659273/

Reviewed By: sophiebits

Differential Revision: D7710224

fbshipit-source-id: 7c673a565f79e45bced91db5b406974f9813327d