Fix bug in merge blocks
Summary:
I read the docs wrong for `Immutable.OrderedMap` set & merge and we don't need to reorder blocks in the `mergeBlocks` operation. We can just update the relevant blocks and delete the empty parent block.

Adds a new test case for the operation that was failing earlier and got fixed here:
{F139809753}

Reviewed By: elboman

Differential Revision: D10100468

fbshipit-source-id: c0e5ad0bf1e3ae091cda2091e57b94d8ee488e45