Fix check for tree blocks
Summary: We were wrongly converting from tree to raw whenever a block had a children array, even if no blocks actually have children. This is bad because when we toggle the GK, we clear the children, but we don't necessarily delete the field altogether (and shouldn't). A tree to raw conversion (or vice versa) is needed only when blocks are actually nested.

Reviewed By: agoder

Differential Revision: D10145467

fbshipit-source-id: 10e489e9c3ba3ae0d16c3da4767cd1400345d92e