Update documentation concerning immutable updates (#1884)
Summary:
**Summary**
Adding some inline examples in the documentation around immutable updates for `EditorState`, `ContentState` and `SelectionState` based on the suggestions of #1445.

Let me know if there's more areas in the documentation that could use similar examples or if some editing needs to be done to what I've added.

**Test Plan**

N/A, but I did view in online markdown editor and it looks ok
Pull Request resolved: https://github.com/facebook/draft-js/pull/1884

Differential Revision: D10188040

fbshipit-source-id: c91d935d3ffdb7aee9ccf826af68c91cf1596a0a