Normalize copyright headers to BSD + patents grant and drop the year (#1964)
Summary:
**Summary**

This PR follows up work started by niveditc on #1860 and suspended because of limitations with tooling. Let's give it another try.

- Remove all copyright years
- Correct copyright headers under `website/` referencing the MIT license
- Inline `All rights reserved.` on the same line of the copyright
- Fixes copyright headers of examples incorrectly using headers for source files

The PR also normalizes copyright headers to BSD with patents grant, as some files were later added with the newest MIT license applied to React and other OSS projects. We should discuss the switch to MIT for Draft.js as well.

**Test Plan**

Manual inspection of files + `yarn lint`
Pull Request resolved: https://github.com/facebook/draft-js/pull/1964

Differential Revision: D13562427

fbshipit-source-id: 6a15d29ff56a6340ee2a0f13f278b207f1761d28